requirements.txt
libsass
django-compressor
django-sass-processor

django-debug-toolbar
django-maintenancemode
django-axes

django-bootstrap4
djangorestframework
djangorestframework_simplejwt
djangorestframework_camel_case
django-cors-headers
django-filter

graphene-django
graphene-django-extras
graphene-django-optimizer
django-graphql-jwt

numpy
markdown
django-markdownx